Overhauled all traps damage, Player Damage, Enemy Damage, Enemy Health, Navmesh
I re-created all the scripts that deal damage to the player so that they are easier to modify.

Added variables that control the players attack damage.

Created a EnemyHealth script that all enemies will use to control their health.

Gave the enemies damage.

Created a Navmesh for the AI to use.

Changed Juan's BarkScript so that it uses tags to destroy walls and deal damage to the player.